move `files` package to `tools.common`
