+++ langworld_db_data/adders/feature_adder.py
@@ -14,12 +14,12 @@ from langworld_db_data.constants.literals import (
     KEY_FOR_VALUE_ID,
     KEY_FOR_VALUE_TYPE,
 )
-from langworld_db_data.tools.files.csv_xls import (
+from langworld_db_data.tools.common.files.csv_xls import (
     read_column_from_csv,
     read_dicts_from_csv,
     write_csv,
 )
-from langworld_db_data.tools.files.txt import remove_extra_space
+from langworld_db_data.tools.common.files.txt import remove_extra_space
 from langworld_db_data.tools.ids.extract import (
     extract_category_id,
     extract_feature_index,
