+++ langworld_db_data/removers/feature_remover.py
@@ -1,6 +1,6 @@
 from langworld_db_data import ObjectWithPaths
 from langworld_db_data.constants.literals import KEY_FOR_ID
-from langworld_db_data.tools.files.csv_xls import (
+from langworld_db_data.tools.common.files.csv_xls import (
     read_column_from_csv,
     read_dicts_from_csv,
     remove_rows_with_given_content_in_lookup_column,
