+++ langworld_db_data/tools/common/files/txt.py
