+++ langworld_db_data/tools/featureprofiles/convert_from_excel.py
@@ -10,14 +10,14 @@ from langworld_db_data.constants.literals import (
     KEY_FOR_VALUE_TYPE,
 )
 from langworld_db_data.constants.paths import CONFIG_DIR
+from langworld_db_data.tools.common.files.csv_xls import read_dicts_from_xls
+from langworld_db_data.tools.common.files.json_toml_yaml import read_json_toml_yaml
 from langworld_db_data.tools.featureprofiles.data_structures import (
     ValueForFeatureProfileDictionary,
 )
 from langworld_db_data.tools.featureprofiles.feature_profile_writer_from_dictionary import (  # noqa E501
     FeatureProfileWriterFromDictionary,
 )
-from langworld_db_data.tools.files.csv_xls import read_dicts_from_xls
-from langworld_db_data.tools.files.json_toml_yaml import read_json_toml_yaml
 
 
 def convert_from_excel(path_to_input_excel: Path) -> Path:
