+++ langworld_db_data/tools/featureprofiles/feature_profile_reader.py
@@ -10,10 +10,10 @@ from langworld_db_data.constants.literals import (
     KEY_FOR_VALUE_TYPE,
 )
 from langworld_db_data.constants.paths import FEATURE_PROFILES_DIR
+from langworld_db_data.tools.common.files.csv_xls import read_dicts_from_csv
 from langworld_db_data.tools.featureprofiles.data_structures import (
     ValueForFeatureProfileDictionary,
 )
-from langworld_db_data.tools.files.csv_xls import read_dicts_from_csv
 
 
 class FeatureProfileReader:
