+++ langworld_db_data/tools/featureprofiles/feature_profile_writer_from_dictionary.py
@@ -2,10 +2,10 @@ from dataclasses import asdict
 from pathlib import Path
 
 from langworld_db_data.constants.literals import KEY_FOR_FEATURE_ID
+from langworld_db_data.tools.common.files.csv_xls import write_csv
 from langworld_db_data.tools.featureprofiles.data_structures import (
     ValueForFeatureProfileDictionary,
 )
-from langworld_db_data.tools.files.csv_xls import write_csv
 
 
 class FeatureProfileWriterFromDictionary:
