+++ langworld_db_data/tools/featureprofiles/value_renamer.py
@@ -8,8 +8,8 @@ from langworld_db_data.constants.literals import (
     KEY_FOR_VALUE_ID,
 )
 from langworld_db_data.constants.paths import FEATURE_PROFILES_DIR, INVENTORIES_DIR
-from langworld_db_data.tools.files.csv_xls import read_dicts_from_csv, write_csv
-from langworld_db_data.tools.files.txt import remove_extra_space
+from langworld_db_data.tools.common.files.csv_xls import read_dicts_from_csv, write_csv
+from langworld_db_data.tools.common.files.txt import remove_extra_space
 
 
 class ValueRenamerError(Exception):
