+++ langworld_db_data/validators/asset_validator.py
@@ -2,7 +2,10 @@ from collections import Counter
 from pathlib import Path
 
 from langworld_db_data.constants.paths import ASSETS_DIR, FILE_WITH_DOCULECTS
-from langworld_db_data.tools.files.csv_xls import read_dicts_from_csv, read_plain_rows_from_csv
+from langworld_db_data.tools.common.files.csv_xls import (
+    read_dicts_from_csv,
+    read_plain_rows_from_csv,
+)
 from langworld_db_data.validators.validator import Validator, ValidatorError
