+++ langworld_db_data/validators/genealogy_validator.py
@@ -7,12 +7,12 @@ from langworld_db_data.constants.paths import (
     FILE_WITH_GENEALOGY_HIERARCHY,
     FILE_WITH_GENEALOGY_NAMES,
 )
-from langworld_db_data.tools.files.csv_xls import (
+from langworld_db_data.tools.common.files.csv_xls import (
     check_csv_for_malformed_rows,
     check_csv_for_repetitions_in_column,
 )
-from langworld_db_data.tools.files.json_toml_yaml import check_yaml_file
-from langworld_db_data.tools.files.txt import read_non_empty_lines_from_txt_file
+from langworld_db_data.tools.common.files.json_toml_yaml import check_yaml_file
+from langworld_db_data.tools.common.files.txt import read_non_empty_lines_from_txt_file
 from langworld_db_data.validators.validator import Validator, ValidatorError
