+++ langworld_db_data/validators/validator.py
@@ -2,7 +2,7 @@ from abc import ABC, abstractmethod
 from pathlib import Path
 
 from langworld_db_data.constants.literals import KEY_FOR_ID
-from langworld_db_data.tools.files.csv_xls import read_column_from_csv
+from langworld_db_data.tools.common.files.csv_xls import read_column_from_csv
 
 
 class ValidatorError(ValueError):
