+++ tests/test_tools_ids_update.py
@@ -1,4 +1,6 @@
-from langworld_db_data.tools.files.csv_xls import remove_rows_with_given_content_in_lookup_column
+from langworld_db_data.tools.common.files.csv_xls import (
+    remove_rows_with_given_content_in_lookup_column,
+)
 from langworld_db_data.tools.ids.update import update_indices_after_given_line_number_if_necessary
